brandwatch_extraction/requirements.txt
requests
pyodbc
python-dotenv
prefect
